+++ src/pages/index.tsx
@@ -1,17 +1,17 @@
 import { useSession } from '@faststore/sdk'
 import { graphql } from 'gatsby'
 import { GatsbySeo, JsonLd } from 'gatsby-plugin-next-seo'
-import Hero from 'src/components/sections/Hero'
-import IncentivesHeader from 'src/components/sections/Incentives/IncentivesHeader'
 import { mark } from 'src/sdk/tests/mark'
 import type { PageProps } from 'gatsby'
 import type { HomePageQueryQuery } from '@generated/graphql'
-import IncentivesMock from 'src/components/sections/Incentives/incentivesMock'
 import 'src/styles/pages/homepage.scss'
 import HomeProductShelf from 'src/components/sections/HomeProductShelf'
+import BlockDesktop from 'src/components/sections/ShelfWithFilter'
 import CommonQuestions from 'src/components/sections/CommonQuestions'
+import MainBanner from 'src/components/sections/MainBanner'
 import BannerMedium from 'src/components/sections/BannerMedium'
 import VideoSection from 'src/components/sections/videosection'
+import selectedTabs from 'src/mocks/bestSellerList.json'
 
 export type Props = PageProps<HomePageQueryQuery>
 
@@ -22,6 +22,7 @@ function Page(props: Props) {
       allContentfulCommonQuestions,
       allContentfulVideoSection,
       allContentfulBannerMedium,
+      allContentfulMainBanner,
     },
     location: { pathname, host },
   } = props
@@ -71,18 +72,14 @@ function Page(props: Props) {
         If needed, wrap your component in a <Section /> component
         (not the HTML tag) before rendering it here.
       */}
-      <Hero
-        title="New Products Available"
-        subtitle="At BaseStore you can shop the best tech of 2022. Enjoy and get 10% off on your first purchase."
-        linkText="See all"
-        link="/"
-        imageSrc="https://storeframework.vtexassets.com/arquivos/ids/190897/Photo.jpg"
-        imageAlt="Quest 2 Controller on a table"
+
+      <MainBanner nodes={allContentfulMainBanner.nodes} />
+
+      <BlockDesktop
+        title="Vendidos"
+        pretitle="Mais"
+        navigattionTabs={selectedTabs.data}
       />
-
-      <IncentivesHeader incentives={IncentivesMock} />
-
-      <HomeProductShelf pretitle="" title="Mais vendidos" />
 
       <VideoSection nodes={allContentfulVideoSection.nodes} />
 
@@ -95,7 +92,7 @@ function Page(props: Props) {
 
       <HomeProductShelf
         pretitle="Disciplina Universitária"
-        title="Em caratér especial"
+        title="Em caráter especial"
       />
 
       <CommonQuestions nodes={allContentfulCommonQuestions.nodes} />
@@ -120,6 +117,22 @@ export const querySSG = graphql`
         }
       }
     }
+    allContentfulMainBanner {
+      nodes {
+        title
+        subtitle
+        imageDesktop {
+          url
+        }
+        imageMobile {
+          url
+        }
+        slug
+        buttonLabel
+        buttonColor
+        buttonTextColor
+      }
+    }
     allContentfulVideoSection {
       nodes {
         video {
@@ -132,6 +145,7 @@ export const querySSG = graphql`
         buttonUrl
         content
         title
+        miniText
       }
     }
     allContentfulBannerMedium {