+++ CHANGELOG.md
@@ -6,10 +6,20 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
 ### Added
 
 ### Fixed
+
+## [0.6.0] - 2022-06-29
+
+### Added
+- **GM-89** - Product Shelf with menu selector.
+- **GM-32** - Main Banner component.
+
+### Fixed
+
+- **GM-267** - Video section styles.
+- **GM-294** - Fix shelf title typo.
 
 ## [0.5.0] - 2022-06-23